Fixed up a couple of serialiser bugs.  - service_type is 16bit.  - missing return statement.  - added a bit more debugging too.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@6189 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/serialiser/rsgxscommentitems.h
@@ -43,7 +43,7 @@ class RsGxsCommentItem : public RsGxsMsgItem
 {
 public:
 
-	RsGxsCommentItem(uint8_t service_type): RsGxsMsgItem(service_type, 
+	RsGxsCommentItem(uint16_t service_type): RsGxsMsgItem(service_type, 
 			RS_PKT_SUBTYPE_GXSCOMMENT_COMMENT_ITEM) {return; }
         virtual ~RsGxsCommentItem() { return;}
         void clear();
@@ -57,7 +57,7 @@ class RsGxsVoteItem : public RsGxsMsgItem
 {
 public:
 
-	RsGxsVoteItem(uint8_t service_type): RsGxsMsgItem(service_type, 
+	RsGxsVoteItem(uint16_t service_type): RsGxsMsgItem(service_type, 
 			RS_PKT_SUBTYPE_GXSCOMMENT_VOTE_ITEM) {return; }
         virtual ~RsGxsVoteItem() { return;}
         void clear();
